backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.121.3",
    "ffmpeg-python>=0.2.0",
    "google-generativeai>=0.8.5",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "static-ffmpeg>=2.13",
    "uvicorn>=0.38.0",
]
